Support style template file input
