+++ style.txt
@@ -1,0 +1,3 @@
+lx,true
+ps,true
+ge,false
